Changed folder structure, added HUD prefab and clock.
