+++ Assets/Story/Prefabs.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: b858d2bab0b6b7f47abd9d81dd368a1b
+folderAsset: yes
+timeCreated: 1525337690
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
