+++ Assets/Story/Prefabs/HUD.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 06cc454c5906f924e8890ee65578aa57
+timeCreated: 1525337678
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
