+++ Assets/Story/Scenes.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6c14e3a81dabb924a8b9d13a21f266b1
+folderAsset: yes
+timeCreated: 1525337721
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
