+++ Assets/Story/Scripts.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 81a4037f6a296b84a9042882b6ba1a47
+folderAsset: yes
+timeCreated: 1525337698
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
