+++ Assets/Story/Scripts/GlobalVariables.cs
