+++ Assets/Story/Scripts/GlobalVariables.cs.meta
