+++ Assets/Story/Scripts/PhrasesScript.cs
